pyproject.toml
[tool.poetry]
name = "story"
version = "0.1.0"
description = ""
authors = ["To Kirill <tio.kirill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
web3 = "6.14.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
httpx = "^0.28.0"
h2 = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
